+++ iothub_client/tests/iothubmessage_ut/iothubmessage_ut.c
@@ -452,7 +452,7 @@ TEST_FUNCTION(IoTHubMessage_CreateFromByteArray_fails)
         umock_c_negative_tests_fail_call(index);
 
         char tmp_msg[64];
-        sprintf(tmp_msg, "IoTHubMessage_CreateFromByteArray failure in test %lu/%lu", index, count);
+        sprintf(tmp_msg, "IoTHubMessage_CreateFromByteArray failure in test %lu/%lu", (unsigned long)index, (unsigned long)count);
 
         IOTHUB_MESSAGE_HANDLE h = IoTHubMessage_CreateFromByteArray(c, 1);
 
@@ -524,7 +524,7 @@ TEST_FUNCTION(IoTHubMessage_CreateFromString_fails)
         umock_c_negative_tests_fail_call(index);
 
         char tmp_msg[64];
-        sprintf(tmp_msg, "IoTHubMessage_CreateFromByteArray failure in test %lu/%lu", index, count);
+        sprintf(tmp_msg, "IoTHubMessage_CreateFromByteArray failure in test %lu/%lu", (unsigned long)index, (unsigned long)count);
 
         IOTHUB_MESSAGE_HANDLE h = IoTHubMessage_CreateFromString("a");
 
@@ -807,7 +807,7 @@ TEST_FUNCTION(IoTHubMessage_Clone_with_BYTE_ARRAY_fails)
         umock_c_negative_tests_fail_call(index);
 
         char tmp_msg[64];
-        sprintf(tmp_msg, "IoTHubMessage_Clone failure in test %lu/%lu", index, count);
+        sprintf(tmp_msg, "IoTHubMessage_Clone failure in test %lu/%lu", (unsigned long)index, (unsigned long)count);
 
         IOTHUB_MESSAGE_HANDLE r = IoTHubMessage_Clone(h);
 
@@ -870,7 +870,7 @@ TEST_FUNCTION(IoTHubMessage_Clone_with_STRING_fails)
         umock_c_negative_tests_fail_call(index);
 
         char tmp_msg[64];
-        sprintf(tmp_msg, "IoTHubMessage_Clone_w_string failure in test %lu/%lu", index, count);
+        sprintf(tmp_msg, "IoTHubMessage_Clone_w_string failure in test %lu/%lu", (unsigned long)index, (unsigned long)count);
 
         IOTHUB_MESSAGE_HANDLE r = IoTHubMessage_Clone(h);
 
@@ -1322,7 +1322,7 @@ TEST_FUNCTION(IoTHubMessage_SetContentTypeSystemProperty_Fails)
         umock_c_negative_tests_fail_call(index);
 
         char tmp_msg[64];
-        sprintf(tmp_msg, "Failed in test %lu/%lu", index, count);
+        sprintf(tmp_msg, "Failed in test %lu/%lu", (unsigned long)index, (unsigned long)count);
 
         //act
         IOTHUB_MESSAGE_RESULT result = IoTHubMessage_SetContentTypeSystemProperty(h, TEST_CONTENT_TYPE);
@@ -1433,7 +1433,7 @@ TEST_FUNCTION(IoTHubMessage_SetContentEncodingSystemProperty_Fails)
         umock_c_negative_tests_fail_call(index);
 
         char tmp_msg[64];
-        sprintf(tmp_msg, "Failed in test %lu/%lu", index, count);
+        sprintf(tmp_msg, "Failed in test %lu/%lu", (unsigned long)index, (unsigned long)count);
 
         //act
         IOTHUB_MESSAGE_RESULT result = IoTHubMessage_SetContentEncodingSystemProperty(h, TEST_CONTENT_ENCODING);
@@ -1703,7 +1703,7 @@ TEST_FUNCTION(IoTHubMessage_SetDiagnosticPropertyData_Fails)
         umock_c_negative_tests_fail_call(index);
 
         char tmp_msg[64];
-        sprintf(tmp_msg, "Failed in test %lu/%lu", index, count);
+        sprintf(tmp_msg, "Failed in test %lu/%lu", (unsigned long)index, (unsigned long)count);
 
         //act
         IOTHUB_MESSAGE_RESULT result = IoTHubMessage_SetDiagnosticPropertyData(h, &TEST_DIAGNOSTIC_DATA);
